+++ src/FSharp.Actor/Transports.fs
@@ -18,11 +18,11 @@ type TCPTransport(config:TcpConfig, ?logger) as self =
     let mutable token = Unchecked.defaultof<CancellationToken>
     let mutable serializer = Unchecked.defaultof<ISerializer>
 
-    let handler =(fun (address:NetAddress, msgId, payload) -> 
+    let handler =(fun (_:NetAddress, _, payload) -> 
                     try
                         receivedRate(1L)
                         let msg = serializer.Deserialize<RemoteMessage>(payload)
-                        Message.postMessage (!!msg.Target) { Sender = (new RemoteActor(msg.Sender, self) :> IActor |> ActorRef); Message = msg.Message; Id = msg.Id }
+                        Message.postMessage msg.Target { Sender = (new RemoteActor(msg.Sender, self) :> IActor |> ActorRef); Message = msg.Message; Id = msg.Id }
                     with e -> 
                         logger.Error("Error handling message: " + e.Message, exn = e)
                  )